pyproject.toml
[tool.poetry]
name = "postgress-da-ai-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "postgres_da_ai_agent"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
openai = ">=1.2,<1.3"
pyautogen = "^0.2.1"
psycopg2-binary = "^2.9.9"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
transformers = "^4.36.1"
torch = "^2.1.2"
scikit-learn = "^1.3.2"
tiktoken = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "postgres_da_ai_agent.main:main"